RN-64 Ability to open the app in offline mode
