+++ app/components/at_mention/index.js
@@ -1,5 +1,17 @@
 // Copyright (c) 2017-present Mattermost, Inc. All Rights Reserved.
 // See License.txt for license information.
 
-import AtMention from './at_mention_container';
-export default AtMention;
+import {connect} from 'react-redux';
+
+import {getUsersByUsername} from 'mattermost-redux/selectors/entities/users';
+
+import AtMention from './at_mention';
+
+function mapStateToProps(state, ownProps) {
+    return {
+        usersByUsername: getUsersByUsername(state),
+        ...ownProps
+    };
+}
+
+export default connect(mapStateToProps)(AtMention);
