+++ app/components/error_list/index.js
@@ -1,6 +1,27 @@
 // Copyright (c) 2017-present Mattermost, Inc. All Rights Reserved.
 // See License.txt for license information.
 
-import ErrorListContainer from './error_list_container';
+import {bindActionCreators} from 'redux';
+import {connect} from 'react-redux';
 
-export default ErrorListContainer;
+import {getDisplayableErrors} from 'mattermost-redux/selectors/errors';
+import {dismissError, clearErrors} from 'mattermost-redux/actions/errors';
+
+import ErrorList from './error_list';
+
+function mapStateToProps(state) {
+    return {
+        errors: getDisplayableErrors(state)
+    };
+}
+
+function mapDispatchToProps(dispatch) {
+    return {
+        actions: bindActionCreators({
+            dismissError,
+            clearErrors
+        }, dispatch)
+    };
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(ErrorList);
