+++ app/components/post_list/index.js
@@ -1,6 +1,17 @@
 // Copyright (c) 2017-present Mattermost, Inc. All Rights Reserved.
 // See License.txt for license information.
 
-import PostList from './post_list_container';
+import {connect} from 'react-redux';
 
-export default PostList;
+import {getTheme} from 'app/selectors/preferences';
+
+import PostList from './post_list';
+
+function mapStateToProps(state, ownProps) {
+    return {
+        theme: getTheme(state),
+        ...ownProps
+    };
+}
+
+export default connect(mapStateToProps)(PostList);
