+++ app/scenes/account_settings/index.js
@@ -1,6 +1,26 @@
 // Copyright (c) 2017-present Mattermost, Inc. All Rights Reserved.
 // See License.txt for license information.
 
-import AccountSettingsContainer from './account_settings_container';
+import {bindActionCreators} from 'redux';
 
-export default AccountSettingsContainer;
+import {goToAccountNotifications} from 'app/actions/navigation';
+import {getTheme} from 'app/selectors/preferences';
+
+import navigationSceneConnect from '../navigationSceneConnect';
+import AccountSettings from './account_settings';
+
+function mapStateToProps(state) {
+    return {
+        theme: getTheme(state)
+    };
+}
+
+function mapDispatchToProps(dispatch) {
+    return {
+        actions: bindActionCreators({
+            goToAccountNotifications
+        }, dispatch)
+    };
+}
+
+export default navigationSceneConnect(mapStateToProps, mapDispatchToProps)(AccountSettings);
