+++ app/scenes/index.js
@@ -13,17 +13,17 @@ import CreateChannel from './create_channel';
 import EditPost from './edit_post';
 import ImagePreview from './image_preview';
 import LoadTeam from './load_team';
-import Login from './login/login_container.js';
+import Login from './login';
 import LoginOptions from './login_options';
 import Mfa from './mfa';
 import MoreChannels from './more_channels';
 import MoreDirectMessages from './more_dms';
 import OptionsModal from './options_modal';
 import Settings from './settings';
-import Root from './root/root_container.js';
-import Search from './search/search_container.js';
-import SelectServer from './select_server/select_server_container.js';
-import SelectTeam from './select_team/select_team_container.js';
+import Root from './root';
+import Search from './search';
+import SelectServer from './select_server';
+import SelectTeam from './select_team';
 import Thread from './thread';
 import UserProfile from './user_profile';
 import Saml from './saml';
